refactor: split event_accumulator.go into topical files (M2-1)
